Support FABDEM file format. (#19)
* Support FABDEM file format.

* Handle NaNs in source data.

* Update README.md

+++ scripts/run_glo_prominence.py
@@ -1,4 +1,5 @@
-# Download GLO tiles and run the prominence program on them
+# Download GLO30 tiles and run the prominence program on them.
+# Also supports FABDEM, which is a variant of GLO30.
 #
 # Requires these programs to be in the path:
 #
@@ -18,7 +19,7 @@ def run_command(command_string):
     retval = subprocess.call(command_string, shell=True)
 
 def process_tile(args):
-    (lat, lng, tile_dir) = args
+    (lat, lng, tile_dir, intermediate_dir, tile_format, download_tiles) = args
     lat_str = "%02d" % abs(lat)
     lng_str = "%03d" % abs(lng)
 
@@ -32,16 +33,21 @@ def process_tile(args):
     else:
         fname_lng = f"W{lng_str}"
 
-        
-    filename_base = f"Copernicus_DSM_COG_10_{fname_lat}_00_{fname_lng}_00_DEM"
 
-    flt_filename = os.path.join(tile_dir, filename_base +  ".flt")
+    if tile_format == "GLO30":
+        filename_base = f"Copernicus_DSM_COG_10_{fname_lat}_00_{fname_lng}_00_DEM"
+    elif tile_format == "FABDEM":
+        filename_base = f"{fname_lat}{fname_lng}_FABDEM_V1-0"
+    else:
+        assert False, f"Unknown tile format {tile_format}"
+
+    flt_filename = os.path.join(intermediate_dir, filename_base +  ".flt")
     
     # FLT tile already exists locally?
     if not os.path.isfile(flt_filename):
         tif_filename = os.path.join(tile_dir, filename_base + ".tif")
-        # Download TIF file if we don't have it yet
-        if not os.path.isfile(tif_filename):
+        # Download TIF file if we don't have it yet?
+        if download_tiles and not os.path.isfile(tif_filename):
             aws_filename = f"s3://copernicus-dem-30m/{filename_base}/{filename_base}.tif"
             aws_command = f"aws s3 cp --no-sign-request {aws_filename} {tile_dir}"
             run_command(aws_command)
@@ -54,7 +60,7 @@ def process_tile(args):
 
             # Delete unneeded auxiliary files
             for extension in [".flt.aux.xml", ".hdr", ".prj"]:
-                filename_to_delete = os.path.join(tile_dir, filename_base + extension)
+                filename_to_delete = os.path.join(intermediate_dir, filename_base + extension)
                 if os.path.isfile(filename_to_delete):
                     os.remove(filename_to_delete)
         
@@ -64,9 +70,17 @@ def main():
     parser = argparse.ArgumentParser(description='Download GLO tiles and run prominence on them')
     requiredNamed = parser.add_argument_group('required named arguments')
     requiredNamed.add_argument('--tile_dir', required = True,
-                              help="Directory to place input tiles")
+                              help="Directory to place source tiles")
     requiredNamed.add_argument('--output_dir', required = True,
                                help="Directory to place prominence results")
+    parser.add_argument('--intermediate_dir',
+                        help="Directory to place unzipped FLT tiles; default=tile_dir")
+    parser.add_argument('--download_tiles', dest='download_tiles', action='store_true',
+                        help="Download any missing tiles")
+    parser.add_argument('--no_download_tiles', dest='download_tiles', action='store_false',
+                        help="Don't download any missing tiles")
+    parser.set_defaults(download_tiles=True)
+
     parser.add_argument('--prominence_command', default='release/prominence',
                         help="Path to prominence binary")
     parser.add_argument('--kml_polygon',
@@ -75,6 +89,8 @@ def main():
                         help="Number of threads to use in computing prominence")
     parser.add_argument('--min_prominence', default=100, type=float,
                         help="Min prominence of resultant peaks, in input units")
+    parser.add_argument('--format', default="GLO30", type=str, choices={"GLO30", "FABDEM"},
+                        help="Format of the input tiles")
     parser.add_argument('min_lat', type=int)
     parser.add_argument('max_lat', type=int)
     parser.add_argument('min_lng', type=int)
@@ -82,6 +98,7 @@ def main():
 
     args = parser.parse_args()
     full_tile_dir = os.path.expanduser(args.tile_dir)
+    intermediate_dir = args.intermediate_dir or full_tile_dir
 
     # Set up filtering polygon
     filterPolygon = Filter()
@@ -102,14 +119,16 @@ def main():
 
             # Skip tiles that don't intersect filtering polygon
             if filterPolygon.intersects(lat, lat + 1, lng, lng + 1):
-                process_args.append((lat, wrappedLng, full_tile_dir))
+                process_args.append((lat, wrappedLng, full_tile_dir, intermediate_dir,
+                                     args.format, args.download_tiles))
 
     pool.map(process_tile, process_args)
     pool.close()
     pool.join()
 
     # Run prominence
-    prom_command = f"{args.prominence_command} --v=1 -f GLO30 -i {args.tile_dir} -o {args.output_dir}" + \
+    prom_command = f"{args.prominence_command} --v=1 -f {args.format}" + \
+        f" -i {args.intermediate_dir} -o {args.output_dir}" + \
         f" -t {args.threads} -m {args.min_prominence}"
 
     if args.kml_polygon:
